+++ platformio/commands/test.py
@@ -70,10 +70,11 @@ def cli(ctx, environment, ignore, upload_port, project_dir, verbose):
             # check ignore patterns
             _ignore = list(ignore)
             if projectconf.has_option(section, "test_ignore"):
-                _ignore.extend([p.strip()
-                                for p in projectconf.get(
-                                    section, "test_ignore").split(",")
-                                if p.strip()])
+                _ignore.extend([
+                    p.strip()
+                    for p in projectconf.get(section, "test_ignore").split(",")
+                    if p.strip()
+                ])
             if testname != "*" and \
                any([fnmatch(testname, p) for p in _ignore]):
                 results.append((None, testname, envname))